@@ -4,38 +4,78 @@ version: 0.0.1
 homepage: https://github.com/gairick-saha/dev_essentials.git
 
 environment:
-  sdk: ">=3.0.1 <4.0.0"
+  sdk: ">=3.2.3 <4.0.0"
   flutter: ">=1.17.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  flutter_hooks: ^0.18.6
+  flutter_web_plugins:
+    sdk: flutter
+
+  flutter_localizations:
+    sdk: flutter
+
+  flutter_hooks: ^0.20.4
   collection: ^1.17.0
-  intl: ^0.18.0
   dio: ^5.1.1
   progress_loading_button: ^2.0.2
   keyboard_dismisser: ^3.0.0
-  package_info_plus: ^4.0.0
+  package_info_plus: ^5.0.1
   device_info_plus: ^9.0.0
   universal_io: ^2.2.0
-  flutter_bloc: ^8.1.2
+  flutter_bloc: ^8.1.3
   mime: ^1.0.4
   flutter_svg: ^2.0.5
   path_provider: ^2.0.14
   path: ^1.8.2
-  bot_toast: ^4.0.3
+  bot_toast: ^4.1.3
   device_preview: ^1.1.0
   sliver_tools: ^0.2.10
-  reactive_forms: ^15.0.0
-  reactive_pin_code_fields: ^1.0.0
+  reactive_forms: ^16.1.0
+  reactive_pin_code_fields: ^2.0.0
   jwt_decoder: ^2.0.1
+  intl: ^0.18.1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
   flutter_lints: ^2.0.0
 
+# For information on the generic Dart part of this file, see the
+# following page: https://dart.dev/tools/pub/pubspec
+
+# The following section is specific to Flutter packages.
 flutter:
-  uses-material-design: true
+
+  # To add assets to your package, add an assets section, like this:
+  # assets:
+  #   - images/a_dot_burr.jpeg
+  #   - images/a_dot_ham.jpeg
+  #
+  # For details regarding assets in packages, see
+  # https://flutter.dev/assets-and-images/#from-packages
+  #
+  # An image asset can refer to one or more resolution-specific "variants", see
+  # https://flutter.dev/assets-and-images/#resolution-aware
+
+  # To add custom fonts to your package, add a fonts section here,
+  # in this "flutter" section. Each entry in this list should have a
+  # "family" key with the font family name, and a "fonts" key with a
+  # list giving the asset and other descriptors for the font. For
+  # example:
+  # fonts:
+  #   - family: Schyler
+  #     fonts:
+  #       - asset: fonts/Schyler-Regular.ttf
+  #       - asset: fonts/Schyler-Italic.ttf
+  #         style: italic
+  #   - family: Trajan Pro
+  #     fonts:
+  #       - asset: fonts/TrajanPro.ttf
+  #       - asset: fonts/TrajanPro_Bold.ttf
+  #         weight: 700
+  #
+  # For details regarding fonts in packages, see
+  # https://flutter.dev/custom-fonts/#from-packages
