@@ -1,56 +1,18 @@
 part of '../hooks.dart';
 
 DevEssentialHookState useDevEssentialHook({
-  required ThemeData theme,
-  required ThemeData darkTheme,
-  required ThemeMode themeMode,
-  required DevEssentialTransition? defaultTransition,
-  required DevEssentialCustomTransition? customTransition,
-  required List<DevEssentialPage>? pages,
-  required Duration? defaultTransitionDuration,
-  required Curve? defaultTransitionCurve,
-  required Curve? defaultDialogTransitionCurve,
-  required Duration? defaultDialogTransitionDuration,
+  required DevEssentialAppConfigData config,
 }) =>
-    use(
-      DevEssentialHook(
-        theme: theme,
-        darkTheme: darkTheme,
-        themeMode: themeMode,
-        defaultTransition: defaultTransition,
-        customTransition: customTransition,
-        defaultTransitionDuration: defaultTransitionDuration,
-        defaultTransitionCurve: defaultTransitionCurve,
-        defaultDialogTransitionCurve: defaultDialogTransitionCurve,
-        defaultDialogTransitionDuration: defaultDialogTransitionDuration,
-        pages: pages,
-      ),
-    );
+    use(DevEssentialHook(
+      config: config,
+    ));
 
 class DevEssentialHook extends Hook<DevEssentialHookState> {
   const DevEssentialHook({
-    required this.theme,
-    required this.darkTheme,
-    required this.themeMode,
-    required this.defaultTransition,
-    required this.customTransition,
-    required this.pages,
-    this.defaultTransitionDuration,
-    this.defaultTransitionCurve,
-    this.defaultDialogTransitionCurve,
-    this.defaultDialogTransitionDuration,
+    required this.config,
   });
 
-  final ThemeData theme;
-  final ThemeData darkTheme;
-  final ThemeMode themeMode;
-  final DevEssentialTransition? defaultTransition;
-  final DevEssentialCustomTransition? customTransition;
-  final Duration? defaultTransitionDuration;
-  final Curve? defaultTransitionCurve;
-  final Curve? defaultDialogTransitionCurve;
-  final Duration? defaultDialogTransitionDuration;
-  final List<DevEssentialPage>? pages;
+  final DevEssentialAppConfigData config;
 
   @override
   HookState<DevEssentialHookState, Hook<DevEssentialHookState>> createState() =>
@@ -58,7 +20,8 @@ class DevEssentialHook extends Hook<DevEssentialHookState> {
 }
 
 class DevEssentialHookState
-    extends HookState<DevEssentialHookState, DevEssentialHook> {
+    extends HookState<DevEssentialHookState, DevEssentialHook>
+    with WidgetsBindingObserver {
   static DevEssentialHookState? _instance;
   static DevEssentialHookState get instance {
     if (_instance == null) {
@@ -68,136 +31,227 @@ class DevEssentialHookState
     }
   }
 
-  late ThemeData _lightTheme;
-  late ThemeData _darkTheme;
-  late ThemeMode _currentThemeMode;
+  late DevEssentialAppConfigData config;
 
-  ThemeData get lightTheme => _lightTheme;
-  ThemeData get darkTheme => _darkTheme;
-  ThemeMode get themeMode => _currentThemeMode;
-
-  ThemeData get theme {
-    if (_currentThemeMode == ThemeMode.dark) {
-      return darkTheme;
-    } else {
-      return lightTheme;
-    }
-  }
-
-  bool defaultPopGesture = Platform.isIOS;
-  bool defaultOpaqueRoute = true;
-
-  late DevEssentialTransition defaultTransition;
-
-  late Duration defaultTransitionDuration = 300.milliseconds;
-  late Curve defaultTransitionCurve = Curves.easeOutQuad;
-
-  late Curve defaultDialogTransitionCurve = Curves.easeOutQuad;
-
-  late Duration defaultDialogTransitionDuration = 300.milliseconds;
-
-  DevEssentialCustomTransition? customTransition;
-
-  final DevEssentialRouting routing = DevEssentialRouting();
-
-  Map<String, String> parameters = {};
-
-  final DevEssentialRoutingTree routingTree = DevEssentialRoutingTree(
-    routes: [],
-  );
-
-  final GlobalKey<NavigatorState> rootNavigatorKey =
-      GlobalKey<NavigatorState>();
-
-  final Map<Object, GlobalKey<NavigatorState>> nestedNavigatorKeys = {};
-
-  GlobalKey<NavigatorState> addNestedNavigatorKey(Object key) {
-    nestedNavigatorKeys.putIfAbsent(
-      key,
-      () => GlobalKey<NavigatorState>(
-        debugLabel: 'DevEssential nested key: ${key.toString()}',
-      ),
-    );
-    return nestedNavigatorKeys[key]!;
-  }
-
-  void _addPages() {
-    if (hook.pages != null) {
-      routingTree.addRoutes(hook.pages!);
-    }
+  @override
+  void dispose() {
+    config.onDispose?.call();
+    Dev.clearTranslations();
+    _instance = null;
+    Engine.removeObserver(this);
+    super.dispose();
   }
 
   @override
   void initHook() {
-    _addPages();
-
-    _lightTheme = hook.theme;
-    _darkTheme = hook.darkTheme;
-    _currentThemeMode = hook.themeMode;
-
-    defaultTransition =
-        hook.defaultTransition ?? DevEssentialTransition.rightToLeft;
-    defaultTransitionDuration =
-        hook.defaultTransitionDuration ?? 300.milliseconds;
-    defaultTransitionCurve = hook.defaultTransitionCurve ?? Curves.easeOutQuad;
-    defaultDialogTransitionCurve =
-        hook.defaultDialogTransitionCurve ?? Curves.easeOutQuad;
-    defaultDialogTransitionDuration =
-        hook.defaultDialogTransitionDuration ?? 300.milliseconds;
-    customTransition = hook.customTransition;
-
+    config = hook.config;
+    Engine.addObserver(this);
     DevEssentialHookState._instance = this;
+    onInit();
     super.initHook();
   }
 
-  @override
-  void didUpdateHook(covariant DevEssentialHook oldHook) {
-    if (oldHook.themeMode != hook.themeMode ||
-        oldHook.theme != hook.theme ||
-        oldHook.darkTheme != hook.darkTheme ||
-        oldHook.defaultTransition != hook.defaultTransition ||
-        oldHook.defaultTransitionDuration != hook.defaultTransitionDuration ||
-        oldHook.defaultTransitionCurve != hook.defaultTransitionCurve ||
-        oldHook.defaultDialogTransitionCurve !=
-            hook.defaultDialogTransitionCurve ||
-        oldHook.defaultDialogTransitionDuration !=
-            hook.defaultDialogTransitionDuration ||
-        oldHook.customTransition != hook.customTransition) {
-      _lightTheme = hook.theme;
-      _darkTheme = hook.darkTheme;
-      _currentThemeMode = hook.themeMode;
-
-      defaultTransition =
-          hook.defaultTransition ?? DevEssentialTransition.rightToLeft;
-      defaultTransitionDuration =
-          hook.defaultTransitionDuration ?? 300.milliseconds;
-      defaultTransitionCurve =
-          hook.defaultTransitionCurve ?? Curves.easeOutQuad;
-      defaultDialogTransitionCurve =
-          hook.defaultDialogTransitionCurve ?? Curves.easeOutQuad;
-      defaultDialogTransitionDuration =
-          hook.defaultDialogTransitionDuration ?? 300.milliseconds;
-      customTransition = hook.customTransition;
+  void onInit() {
+    if (config.unknownRoute == null) {
+      config =
+          config.copyWith(unknownRoute: DevEssentialPages.defaultUnknownRoute);
     }
 
-    DevEssentialHookState._instance = this;
-    setState(() {});
-    super.didUpdateHook(oldHook);
-  }
+    if (config.navigatorObservers == null) {
+      config = config.copyWith(
+        navigatorObservers: <NavigatorObserver>[
+          DevEssentialNavigationObserver(
+            config.routingCallback,
+            Dev.routing,
+          ),
+          BotToastNavigatorObserver(),
+        ],
+      );
+    } else {
+      config = config.copyWith(
+        navigatorObservers: <NavigatorObserver>[
+          DevEssentialNavigationObserver(
+            config.routingCallback,
+            config.routing,
+          ),
+          BotToastNavigatorObserver(),
+          ...config.navigatorObservers!
+        ],
+      );
+    }
 
-  void changeThemeData(ThemeData themeData) => setState(
-        () {
-          if (_currentThemeMode == ThemeMode.dark) {
-            _darkTheme = themeData;
-          } else {
-            _lightTheme = themeData;
-          }
-        },
+    if (config.routerDelegate == null) {
+      final newDelegate = DevEssentialRouterDelegate.createDelegate(
+        pages: DevEssentialPages.getPages(
+          config.pages ??
+              [
+                DevEssentialPage(
+                  name: cleanRouteName("/${config.home.runtimeType}"),
+                  page: (_, __) =>
+                      config.home ??
+                      const Scaffold(
+                        body: Center(
+                          child: Text('No pages of home provided'),
+                        ),
+                      ),
+                ),
+              ],
+          splashConfig: config.splashConfig,
+        ),
+        notFoundRoute: config.unknownRoute,
+        navigatorKey: config.navigatorKey,
+        navigatorObservers: config.navigatorObservers,
       );
 
-  void changeThemeMode(ThemeMode themeMode) =>
-      setState(() => _currentThemeMode = themeMode);
+      config = config.copyWith(routerDelegate: newDelegate);
+
+      if (config.routeInformationParser == null) {
+        final newRouteInformationParser =
+            DevEssentialRouteInformationParser.createInformationParser(
+          initialRoute: config.initialRoute ??
+              config.pages?.first.name ??
+              cleanRouteName("/${config.home.runtimeType}"),
+        );
+
+        config =
+            config.copyWith(routeInformationParser: newRouteInformationParser);
+      }
+
+      if (config.locale != null) Dev.locale = config.locale;
+
+      if (config.fallbackLocale != null) {
+        Dev.fallbackLocale = config.fallbackLocale;
+      }
+
+      if (config.translations != null) {
+        Dev.addTranslations(config.translations!.keys);
+      } else if (config.translationsKeys != null) {
+        Dev.addTranslations(config.translationsKeys!);
+      }
+
+      config.onInit?.call();
+
+      Dev.isLogEnable = config.enableLog ?? kDebugMode;
+      Dev.log = config.logWriterCallback ?? defaultLogWriterCallback;
+
+      Engine.schedulerBinding.addPostFrameCallback((_) {
+        if (config.defaultTransition == null) {
+          config = config.copyWith(defaultTransition: getThemeTransition());
+        }
+      });
+
+      onReady();
+    }
+  }
+
+  DevEssentialTransition? getThemeTransition() {
+    final platform = context.theme.platform;
+    final matchingTransition =
+        Theme.of(context).pageTransitionsTheme.builders[platform];
+    switch (matchingTransition) {
+      case CupertinoPageTransitionsBuilder():
+        return DevEssentialTransition.cupertino;
+      case ZoomPageTransitionsBuilder():
+        return DevEssentialTransition.zoom;
+      case FadeUpwardsPageTransitionsBuilder():
+        return DevEssentialTransition.fade;
+      case OpenUpwardsPageTransitionsBuilder():
+        return DevEssentialTransition.native;
+      default:
+        return DevEssentialTransition.native;
+    }
+  }
+
+  void onReady() => config.onReady?.call();
+
+  set parameters(Map<String, String?> newParameters) =>
+      config = config.copyWith(parameters: newParameters);
+
+  set testMode(bool isTest) => config = config.copyWith(testMode: isTest);
 
   @override
-  DevEssentialHookState build(BuildContext context) => this;
+  void didChangeLocales(List<Locale>? locales) {
+    Dev.asap(() {
+      final locale = Dev.deviceLocale;
+      if (locale != null) {
+        Dev.updateLocale(locale);
+      }
+    });
+  }
+
+  void setTheme(ThemeData value) {
+    if (config.darkTheme == null) {
+      config = config.copyWith(theme: value);
+    } else {
+      if (value.brightness == Brightness.light) {
+        config = config.copyWith(theme: value);
+      } else {
+        config = config.copyWith(darkTheme: value);
+      }
+    }
+    update();
+  }
+
+  void setThemeMode(ThemeMode value) {
+    config = config.copyWith(themeMode: value);
+    update();
+  }
+
+  void restartApp() {
+    config = config.copyWith(unikey: UniqueKey());
+    update();
+  }
+
+  void update() {
+    context.visitAncestorElements((element) {
+      element.markNeedsBuild();
+      return false;
+    });
+  }
+
+  DevEssentialRouterDelegate get rootDelegate =>
+      config.routerDelegate as DevEssentialRouterDelegate;
+
+  RouteInformationParser<Object> get informationParser =>
+      config.routeInformationParser!;
+
+  GlobalKey<NavigatorState> get key => rootDelegate.navigatorKey;
+
+  GlobalKey<NavigatorState>? addKey(GlobalKey<NavigatorState> newKey) {
+    rootDelegate.navigatorKey = newKey;
+    return key;
+  }
+
+  Map<String, DevEssentialRouterDelegate> keys = {};
+
+  DevEssentialRouterDelegate? nestedKey(String? key) {
+    if (key == null) {
+      return rootDelegate;
+    }
+    keys.putIfAbsent(
+      key,
+      () => DevEssentialRouterDelegate(
+        showHashOnUrl: true,
+        pages: DevEssentialRouteDecoder.fromRoute(key).currentChildren ?? [],
+        notFoundRoute: config.unknownRoute,
+        navigatorKey: GlobalKey<NavigatorState>(debugLabel: key),
+        navigatorObservers: config.navigatorObservers,
+      ),
+    );
+    return keys[key];
+  }
+
+  @override
+  DevEssentialHookState build(BuildContext context) {
+    return this;
+  }
+
+  String cleanRouteName(String name) {
+    name = name.replaceAll('() => ', '');
+
+    if (!name.startsWith('/')) {
+      name = '/$name';
+    }
+    return Uri.tryParse(name)?.toString() ?? name;
+  }
 }
