@@ -1,157 +1,205 @@
-part of '../dev_snackbar.dart';
+part of '../snackbar.dart';
 
-class DevEssentialSnackbar extends HookWidget {
-  DevEssentialSnackbar({
+class DevEssentialSnackBar extends StatefulWidget {
+  final SnackbarStatusCallback? snackbarStatus;
+
+  final String? title;
+
+  final DismissDirection? dismissDirection;
+
+  final String? message;
+
+  final Widget? titleText;
+
+  final Widget? messageText;
+
+  final Color backgroundColor;
+
+  final Color? leftBarIndicatorColor;
+
+  final List<BoxShadow>? boxShadows;
+
+  final Gradient? backgroundGradient;
+
+  final Widget? icon;
+
+  final bool shouldIconPulse;
+
+  final Widget? mainButton;
+
+  final OnSnackbarTap? onTap;
+
+  final OnSnackbarHover? onHover;
+
+  final Duration? duration;
+
+  final bool showProgressIndicator;
+
+  final AnimationController? progressIndicatorController;
+
+  final Color? progressIndicatorBackgroundColor;
+
+  final Animation<Color>? progressIndicatorValueColor;
+
+  final bool isDismissible;
+
+  final double? maxWidth;
+
+  final EdgeInsets margin;
+
+  final EdgeInsets padding;
+
+  final double borderRadius;
+
+  final Color? borderColor;
+
+  final double? borderWidth;
+
+  final SnackPosition snackPosition;
+
+  final SnackStyle snackStyle;
+
+  final Curve forwardAnimationCurve;
+
+  final Curve reverseAnimationCurve;
+
+  final Duration animationDuration;
+
+  final double barBlur;
+
+  final double overlayBlur;
+
+  final Color? overlayColor;
+
+  final Form? userInputForm;
+
+  const DevEssentialSnackBar({
     Key? key,
     this.title,
     this.message,
     this.titleText,
     this.messageText,
     this.icon,
-    this.useAnimatedIcon = true,
+    this.shouldIconPulse = true,
     this.maxWidth,
-    this.margin = EdgeInsets.zero,
+    this.margin = const EdgeInsets.all(0.0),
     this.padding = const EdgeInsets.all(16),
-    this.borderRadius = 8.0,
+    this.borderRadius = 0.0,
     this.borderColor,
     this.borderWidth = 1.0,
     this.backgroundColor = const Color(0xFF303030),
-    this.isDismissible = true,
-    this.duration,
-    this.mainButton,
-    this.backgroundGradient,
+    this.leftBarIndicatorColor,
     this.boxShadows,
+    this.backgroundGradient,
+    this.mainButton,
+    this.onTap,
+    this.onHover,
+    this.duration,
+    this.isDismissible = true,
+    this.dismissDirection,
     this.showProgressIndicator = false,
     this.progressIndicatorController,
     this.progressIndicatorBackgroundColor,
     this.progressIndicatorValueColor,
-    this.leftBarIndicatorColor,
-    this.blurValue = 0.0,
-    this.overlayBlurValue = 0.0,
-    this.overlayColor = Colors.transparent,
-    this.userInputForm,
-    this.position = DevEssentialSnackbarPosition.BOTTOM,
-    this.style = DevEssentialSnackbarStyle.FLOATING,
-    this.status,
+    this.snackPosition = SnackPosition.bottom,
+    this.snackStyle = SnackStyle.floating,
     this.forwardAnimationCurve = Curves.easeOutCirc,
     this.reverseAnimationCurve = Curves.easeOutCirc,
     this.animationDuration = const Duration(seconds: 1),
-    this.onTap,
-    this.dismissDirection,
-  })  : assert(
-          message != null || (messageText != null && messageText.isNotEmpty),
-          '''You need to either use messageText[String], or message[Widget] in DevEssentialSnackbar''',
-        ),
-        super(key: key);
+    this.barBlur = 0.0,
+    this.overlayBlur = 0.0,
+    this.overlayColor = Colors.transparent,
+    this.userInputForm,
+    this.snackbarStatus,
+  }) : super(key: key);
 
-  final Widget? title;
-  final String? titleText;
-  final Widget? message;
-  final String? messageText;
-  final Widget? icon;
-  final bool useAnimatedIcon;
-  final double? maxWidth;
-  final EdgeInsets margin;
-  final EdgeInsets padding;
-  final double borderRadius;
-  final Color? borderColor;
-  final double borderWidth;
-  final Color backgroundColor;
-  final bool isDismissible;
-  final Duration? duration;
-  final Widget? mainButton;
-  final Gradient? backgroundGradient;
-  final List<BoxShadow>? boxShadows;
-  final bool showProgressIndicator;
-  final AnimationController? progressIndicatorController;
-  final Color? progressIndicatorBackgroundColor;
-  final Animation<Color>? progressIndicatorValueColor;
-  final Color? leftBarIndicatorColor;
+  @override
+  State createState() => DevEssentialSnackBarState();
 
-  /// Default is 0.0. If different than 0.0, blurs only Snack's background.
-  /// To take effect, make sure your [backgroundColor] has some opacity.
-  /// The greater the value, the greater the blur.
-  final double blurValue;
+  DevEssentialSnackbarController show() => Dev.showSnackbar(this);
+}
 
-  /// Default is 0.0. If different than 0.0, creates a blurred
-  /// overlay that prevents the user from interacting with the screen.
-  /// The greater the value, the greater the blur.
-  final double overlayBlurValue;
+class DevEssentialSnackBarState extends State<DevEssentialSnackBar>
+    with TickerProviderStateMixin {
+  AnimationController? _fadeController;
+  late Animation<double> _fadeAnimation;
 
-  /// Default is [Colors.transparent]. Only takes effect if [overlayBlurValue] > 0.0.
-  /// Make sure you use a color with transparency here e.g.
-  /// Colors.grey[600].withOpacity(0.2).
-  final Color? overlayColor;
+  final Widget _emptyWidget = const SizedBox(width: 0.0, height: 0.0);
+  final double _initialOpacity = 1.0;
+  final double _finalOpacity = 0.4;
 
-  final DevEssentialFormBuilder? userInputForm;
-  final DevEssentialSnackbarPosition position;
-  final DevEssentialSnackbarStyle style;
-  final DevEssentialSnackbarStatusCallback? status;
-  final Curve forwardAnimationCurve;
-  final Curve reverseAnimationCurve;
-  final Duration animationDuration;
+  final Duration _pulseAnimationDuration = const Duration(seconds: 1);
 
-  /// A callback that registers the user's click anywhere.
-  /// An alternative to [mainButton]
-  final DevEssentialSnackbarOnTapCallback? onTap;
+  late bool _isTitlePresent;
+  late double _messageTopMargin;
 
-  /// The direction in which the SnackBar can be dismissed.
-  ///
-  /// Default is [DismissDirection.down] when
-  /// [position] == [DevEssentialSnackbarPosition.BOTTOM] and [DismissDirection.up]
-  /// when [position] == [DevEssentialSnackbarPosition.TOP]
-  final DismissDirection? dismissDirection;
+  FocusScopeNode? _focusNode;
+  late FocusAttachment _focusAttachment;
+
+  final Completer<Size> _boxHeightCompleter = Completer<Size>();
+
+  late CurvedAnimation _progressAnimation;
+
+  final _backgroundBoxKey = GlobalKey();
+
+  double get buttonPadding {
+    if (widget.padding.right - 12 < 0) {
+      return 4;
+    } else {
+      return widget.padding.right - 12;
+    }
+  }
+
+  RowStyle get _rowStyle {
+    if (widget.mainButton != null && widget.icon == null) {
+      return RowStyle.action;
+    } else if (widget.mainButton == null && widget.icon != null) {
+      return RowStyle.icon;
+    } else if (widget.mainButton != null && widget.icon != null) {
+      return RowStyle.all;
+    } else {
+      return RowStyle.none;
+    }
+  }
 
   @override
   Widget build(BuildContext context) {
-    DevEssentialSnackbarHookState snackbarState = useDevEssentialSnackbarHook(
-      progressIndicatorController: progressIndicatorController,
-      showProgressIndicator: showProgressIndicator,
-      configureIconAnimation: icon != null && icon is Icon && useAnimatedIcon,
-      iconAnimationDuration: 1.seconds,
-      iconAnimationFinalOpacity: 0.5,
-      iconAnimationInitialOpacity: 1.0,
-      hasTitle: title != null || titleText != null,
-      padding: padding,
-      margin: margin,
-    );
     return Align(
       heightFactor: 1.0,
       child: Material(
-        color: style == DevEssentialSnackbarStyle.FLOATING
+        color: widget.snackStyle == SnackStyle.floating
             ? Colors.transparent
-            : backgroundColor,
+            : widget.backgroundColor,
         child: SafeArea(
-          minimum: position == DevEssentialSnackbarPosition.BOTTOM
-              ? EdgeInsets.only(bottom: Dev.viewInsets.bottom)
-              : EdgeInsets.only(top: Dev.mediaQuery.padding.top),
-          bottom: position == DevEssentialSnackbarPosition.BOTTOM,
-          top: position == DevEssentialSnackbarPosition.TOP,
+          minimum: widget.snackPosition == SnackPosition.bottom
+              ? EdgeInsets.only(
+                  bottom: MediaQuery.of(context).viewInsets.bottom)
+              : EdgeInsets.only(top: MediaQuery.of(context).padding.top),
+          bottom: widget.snackPosition == SnackPosition.bottom,
+          top: widget.snackPosition == SnackPosition.top,
           left: false,
           right: false,
           child: Stack(
-            clipBehavior: Clip.antiAliasWithSaveLayer,
             children: [
               FutureBuilder<Size>(
-                future: snackbarState._boxHeightCompleter.future,
+                future: _boxHeightCompleter.future,
                 builder: (context, snapshot) {
                   if (snapshot.hasData) {
-                    if (blurValue == 0) {
+                    if (widget.barBlur == 0) {
                       return _emptyWidget;
                     }
                     return ClipRRect(
-                      borderRadius: BorderRadius.circular(borderRadius),
+                      borderRadius: BorderRadius.circular(widget.borderRadius),
                       child: BackdropFilter(
                         filter: ImageFilter.blur(
-                          sigmaX: blurValue,
-                          sigmaY: blurValue,
-                        ),
+                            sigmaX: widget.barBlur, sigmaY: widget.barBlur),
                         child: Container(
                           height: snapshot.data!.height,
                           width: snapshot.data!.width,
                           decoration: BoxDecoration(
                             color: Colors.transparent,
-                            borderRadius: BorderRadius.circular(borderRadius),
+                            borderRadius:
+                                BorderRadius.circular(widget.borderRadius),
                           ),
                         ),
                       ),
@@ -161,10 +209,10 @@ class DevEssentialSnackbar extends HookWidget {
                   }
                 },
               ),
-              if (userInputForm == null)
-                _containerWithoutForm(context, snackbarState)
+              if (widget.userInputForm != null)
+                _containerWithForm()
               else
-                _containerWithForm(context, snackbarState),
+                _containerWithoutForm()
             ],
           ),
         ),
@@ -172,55 +220,50 @@ class DevEssentialSnackbar extends HookWidget {
     );
   }
 
-  double get buttonPadding {
-    if (padding.right - 12 < 0) {
-      return 4;
-    } else {
-      return padding.right - 12;
-    }
+  @override
+  void dispose() {
+    _fadeController?.dispose();
+    widget.progressIndicatorController?.removeListener(_updateProgress);
+    widget.progressIndicatorController?.dispose();
+
+    _focusAttachment.detach();
+    _focusNode!.dispose();
+    super.dispose();
   }
 
-  DevEssentialSnackbarRowStyle get _rowStyle {
-    if (mainButton != null && icon == null) {
-      return DevEssentialSnackbarRowStyle.action;
-    } else if (mainButton == null && icon != null) {
-      return DevEssentialSnackbarRowStyle.icon;
-    } else if (mainButton != null && icon != null) {
-      return DevEssentialSnackbarRowStyle.all;
-    } else {
-      return DevEssentialSnackbarRowStyle.none;
-    }
-  }
+  @override
+  void initState() {
+    super.initState();
 
-  final Widget _emptyWidget = const SizedBox.shrink();
+    assert(
+      widget.userInputForm != null ||
+          ((widget.message != null && widget.message!.isNotEmpty) ||
+              widget.messageText != null),
+      '''You need to either use message[String], or messageText[Widget] or define a userInputForm[Form] in DevEssentialSnackbar''',
+    );
+    _isTitlePresent = (widget.title != null || widget.titleText != null);
+    _messageTopMargin = _isTitlePresent ? 6.0 : widget.padding.top;
 
-  Widget? _getIcon(
-    BuildContext context,
-    DevEssentialSnackbarHookState snackbarState,
-  ) {
-    if (icon != null && icon is Icon && useAnimatedIcon) {
-      return FadeTransition(
-        opacity: snackbarState._iconFadeAnimation,
-        child: icon,
-      );
-    } else if (icon != null) {
-      return icon;
+    _configureLeftBarFuture();
+    _configureProgressIndicatorAnimation();
+
+    if (widget.icon != null && widget.shouldIconPulse) {
+      _configurePulseAnimation();
+      _fadeController?.forward();
     }
 
-    return null;
+    _focusNode = FocusScopeNode();
+    _focusAttachment = _focusNode!.attach(context);
   }
 
-  Widget _buildLeftBarIndicator(
-    BuildContext context,
-    DevEssentialSnackbarHookState snackbarState,
-  ) {
-    if (leftBarIndicatorColor != null) {
+  Widget _buildLeftBarIndicator() {
+    if (widget.leftBarIndicatorColor != null) {
       return FutureBuilder<Size>(
-        future: snackbarState._boxHeightCompleter.future,
+        future: _boxHeightCompleter.future,
         builder: (buildContext, snapshot) {
           if (snapshot.hasData) {
             return Container(
-              color: leftBarIndicatorColor,
+              color: widget.leftBarIndicatorColor,
               width: 5.0,
               height: snapshot.data!.height,
             );
@@ -234,56 +277,123 @@ class DevEssentialSnackbar extends HookWidget {
     }
   }
 
-  Widget _containerWithoutForm(
-    BuildContext context,
-    DevEssentialSnackbarHookState snackbarState,
-  ) {
-    final iconPadding = padding.left > 16.0 ? padding.left : 0.0;
-    final left = _rowStyle == DevEssentialSnackbarRowStyle.icon ||
-            _rowStyle == DevEssentialSnackbarRowStyle.all
-        ? 4.0
-        : padding.left;
-    final right = _rowStyle == DevEssentialSnackbarRowStyle.action ||
-            _rowStyle == DevEssentialSnackbarRowStyle.all
-        ? 8.0
-        : padding.right;
+  void _configureLeftBarFuture() {
+    Engine.instance.addPostFrameCallback(
+      (_) {
+        final keyContext = _backgroundBoxKey.currentContext;
+        if (keyContext != null) {
+          final box = keyContext.findRenderObject() as RenderBox;
+          _boxHeightCompleter.complete(box.size);
+        }
+      },
+    );
+  }
 
-    return AnimatedContainer(
-      duration: kThemeAnimationDuration,
-      constraints:
-          maxWidth != null ? BoxConstraints(maxWidth: maxWidth!) : null,
+  void _configureProgressIndicatorAnimation() {
+    if (widget.showProgressIndicator &&
+        widget.progressIndicatorController != null) {
+      widget.progressIndicatorController!.addListener(_updateProgress);
+
+      _progressAnimation = CurvedAnimation(
+          curve: Curves.linear, parent: widget.progressIndicatorController!);
+    }
+  }
+
+  void _configurePulseAnimation() {
+    _fadeController =
+        AnimationController(vsync: this, duration: _pulseAnimationDuration);
+    _fadeAnimation = Tween(begin: _initialOpacity, end: _finalOpacity).animate(
+      CurvedAnimation(
+        parent: _fadeController!,
+        curve: Curves.linear,
+      ),
+    );
+
+    _fadeController!.addStatusListener((status) {
+      if (status == AnimationStatus.completed) {
+        _fadeController!.reverse();
+      }
+      if (status == AnimationStatus.dismissed) {
+        _fadeController!.forward();
+      }
+    });
+
+    _fadeController!.forward();
+  }
+
+  Widget _containerWithForm() {
+    return Container(
+      key: _backgroundBoxKey,
+      constraints: widget.maxWidth != null
+          ? BoxConstraints(maxWidth: widget.maxWidth!)
+          : null,
       decoration: BoxDecoration(
-        color: backgroundColor,
-        gradient: backgroundGradient,
-        boxShadow: boxShadows,
-        borderRadius: BorderRadius.circular(borderRadius),
-        border: borderColor != null
-            ? Border.all(color: borderColor!, width: borderWidth)
+        color: widget.backgroundColor,
+        gradient: widget.backgroundGradient,
+        boxShadow: widget.boxShadows,
+        borderRadius: BorderRadius.circular(widget.borderRadius),
+        border: widget.borderColor != null
+            ? Border.all(
+                color: widget.borderColor!,
+                width: widget.borderWidth!,
+              )
+            : null,
+      ),
+      child: Padding(
+        padding: const EdgeInsets.only(
+            left: 8.0, right: 8.0, bottom: 8.0, top: 16.0),
+        child: FocusScope(
+          node: _focusNode,
+          autofocus: true,
+          child: widget.userInputForm!,
+        ),
+      ),
+    );
+  }
+
+  Widget _containerWithoutForm() {
+    final iconPadding = widget.padding.left > 16.0 ? widget.padding.left : 0.0;
+    final left = _rowStyle == RowStyle.icon || _rowStyle == RowStyle.all
+        ? 4.0
+        : widget.padding.left;
+    final right = _rowStyle == RowStyle.action || _rowStyle == RowStyle.all
+        ? 8.0
+        : widget.padding.right;
+    return Container(
+      key: _backgroundBoxKey,
+      constraints: widget.maxWidth != null
+          ? BoxConstraints(maxWidth: widget.maxWidth!)
+          : null,
+      decoration: BoxDecoration(
+        color: widget.backgroundColor,
+        gradient: widget.backgroundGradient,
+        boxShadow: widget.boxShadows,
+        borderRadius: BorderRadius.circular(widget.borderRadius),
+        border: widget.borderColor != null
+            ? Border.all(color: widget.borderColor!, width: widget.borderWidth!)
             : null,
       ),
       child: Column(
         mainAxisSize: MainAxisSize.min,
         children: [
-          showProgressIndicator
-              ? LoadingIndictor(
-                  materialLoadingValue: progressIndicatorController != null
-                      ? snackbarState._progressAnimation.value
+          widget.showProgressIndicator
+              ? LinearProgressIndicator(
+                  value: widget.progressIndicatorController != null
+                      ? _progressAnimation.value
                       : null,
-                  materialLoadingBackgroundColor:
-                      progressIndicatorBackgroundColor,
-                  materialLoadingValueColor: progressIndicatorValueColor,
+                  backgroundColor: widget.progressIndicatorBackgroundColor,
+                  valueColor: widget.progressIndicatorValueColor,
                 )
               : _emptyWidget,
           Row(
             mainAxisSize: MainAxisSize.max,
             children: [
-              _buildLeftBarIndicator(context, snackbarState),
-              if (_rowStyle == DevEssentialSnackbarRowStyle.icon ||
-                  _rowStyle == DevEssentialSnackbarRowStyle.all)
+              _buildLeftBarIndicator(),
+              if (_rowStyle == RowStyle.icon || _rowStyle == RowStyle.all)
                 ConstrainedBox(
                   constraints:
                       BoxConstraints.tightFor(width: 42.0 + iconPadding),
-                  child: _getIcon(context, snackbarState),
+                  child: _getIcon(),
                 ),
               Expanded(
                 flex: 1,
@@ -291,55 +401,46 @@ class DevEssentialSnackbar extends HookWidget {
                   crossAxisAlignment: CrossAxisAlignment.stretch,
                   mainAxisSize: MainAxisSize.min,
                   children: <Widget>[
-                    if (snackbarState._isTitlePresent)
+                    if (_isTitlePresent)
                       Padding(
                         padding: EdgeInsets.only(
-                          top: padding.top,
+                          top: widget.padding.top,
                           left: left,
                           right: right,
                         ),
-                        child: title ??
+                        child: widget.titleText ??
                             Text(
-                              titleText ?? "",
-                              textScaleFactor: Dev.textScaleFactor,
-                              style: Theme.of(context)
-                                  .textTheme
-                                  .titleMedium
-                                  ?.copyWith(
-                                    fontSize: 16.0,
-                                    color: Colors.white,
-                                    fontWeight: FontWeight.bold,
-                                  ),
+                              widget.title ?? "",
+                              style: const TextStyle(
+                                fontSize: 16.0,
+                                color: Colors.white,
+                                fontWeight: FontWeight.bold,
+                              ),
                             ),
-                      ),
+                      )
+                    else
+                      _emptyWidget,
                     Padding(
                       padding: EdgeInsets.only(
-                        top: snackbarState._messageTopMargin,
+                        top: _messageTopMargin,
                         left: left,
                         right: right,
-                        bottom: padding.bottom,
+                        bottom: widget.padding.bottom,
                       ),
-                      child: message ??
+                      child: widget.messageText ??
                           Text(
-                            messageText ?? "",
-                            textScaleFactor: Dev.textScaleFactor,
-                            style: Theme.of(context)
-                                .textTheme
-                                .bodyMedium
-                                ?.copyWith(
-                                  fontSize: 14.0,
-                                  color: Colors.white,
-                                ),
+                            widget.message ?? "",
+                            style: const TextStyle(
+                                fontSize: 14.0, color: Colors.white),
                           ),
                     ),
                   ],
                 ),
               ),
-              if (_rowStyle == DevEssentialSnackbarRowStyle.action ||
-                  _rowStyle == DevEssentialSnackbarRowStyle.all)
+              if (_rowStyle == RowStyle.action || _rowStyle == RowStyle.all)
                 Padding(
                   padding: EdgeInsets.only(right: buttonPadding),
-                  child: mainButton,
+                  child: widget.mainButton,
                 ),
             ],
           ),
@@ -348,35 +449,18 @@ class DevEssentialSnackbar extends HookWidget {
     );
   }
 
-  Widget _containerWithForm(
-    BuildContext context,
-    DevEssentialSnackbarHookState snackbarState,
-  ) {
-    return Container(
-      key: snackbarState._backgroundBoxKey,
-      constraints:
-          maxWidth != null ? BoxConstraints(maxWidth: maxWidth!) : null,
-      decoration: BoxDecoration(
-        color: backgroundColor,
-        gradient: backgroundGradient,
-        boxShadow: boxShadows,
-        borderRadius: BorderRadius.circular(borderRadius),
-        border: borderColor != null
-            ? Border.all(
-                color: borderColor!,
-                width: borderWidth,
-              )
-            : null,
-      ),
-      child: Padding(
-        padding: const EdgeInsets.only(
-            left: 8.0, right: 8.0, bottom: 8.0, top: 16.0),
-        child: FocusScope(
-          node: snackbarState._userInputFormFocusNode,
-          autofocus: true,
-          child: userInputForm!,
-        ),
-      ),
-    );
+  Widget? _getIcon() {
+    if (widget.icon != null && widget.icon is Icon && widget.shouldIconPulse) {
+      return FadeTransition(
+        opacity: _fadeAnimation,
+        child: widget.icon,
+      );
+    } else if (widget.icon != null) {
+      return widget.icon;
+    } else {
+      return _emptyWidget;
+    }
   }
+
+  void _updateProgress() => setState(() {});
 }
