@@ -1,14 +1,22 @@
 library dev_essentials;
 
-export 'src/animations/dev_animations.dart';
-export 'src/apps/apps.dart';
+export 'package:package_info_plus/package_info_plus.dart';
+export 'package:device_info_plus/device_info_plus.dart';
+
+export 'src/platform/platform.dart';
 export 'src/clippers/clippers.dart';
-export 'src/configs/configs.dart';
-export 'src/dev_essential.dart';
-export 'src/extensions/extensions.dart';
-export 'src/navigation/navigation.dart';
-export 'src/network/network.dart';
-export 'src/routes/routes.dart';
 export 'src/transitions/transitions.dart';
-export 'src/utils/utils.dart';
+export 'src/translation/translation.dart';
+export 'src/extensions/extensions.dart';
+export 'src/network/network.dart';
+export 'src/utilis/utils.dart';
 export 'src/widgets/widgets.dart';
+export 'src/dev_essential.dart';
+export 'src/url_strategy/url_strategy.dart';
+export 'src/configs/configs.dart';
+export 'src/navigation/navigation.dart';
+export 'src/apps/apps.dart';
+export 'src/snackbar/snackbar.dart';
+export 'src/routes/routes.dart';
+
+export 'src/animations/dev_animations.dart';
