@@ -1,5 +1,7 @@
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
-import '../utils/utils.dart';
+import '../../dev_essentials.dart';
 
+part 'src/hook_config.dart';
 part 'src/splash_config.dart';
