@@ -1,31 +1,44 @@
 part of '../extensions.dart';
 
-extension OverlayExtension on DevEssential {
+extension OverlayExt on DevEssential {
   Future<T> showOverlay<T>({
     required Future<T> Function() asyncFunction,
     Color opacityColor = Colors.black,
     Widget? loadingWidget,
     double opacity = .5,
+    GlobalKey<NavigatorState>? navigatorKey,
   }) async {
-    final navigatorState = Navigator.of(overlayContext!, rootNavigator: false);
+    final navigatorState = navigatorKey?.currentState ??
+        Navigator.of(Dev.overlayContext!, rootNavigator: false);
+
     final overlayState = navigatorState.overlay!;
 
-    final overlayEntryOpacity = OverlayEntry(builder: (context) {
-      return Opacity(
+    final overlayEntryOpacity = OverlayEntry(
+      builder: (context) {
+        return Opacity(
           opacity: opacity,
           child: Container(
             color: opacityColor,
-          ));
-    });
-    final overlayEntryLoader = OverlayEntry(builder: (context) {
-      return loadingWidget ??
-          const Center(
-              child: SizedBox(
-            height: 90,
-            width: 90,
-            child: Text('Loading...'),
-          ));
-    });
+          ),
+        );
+      },
+    );
+    final overlayEntryLoader = OverlayEntry(
+      builder: (context) {
+        return loadingWidget ??
+            Center(
+              child: Material(
+                shape: RoundedRectangleBorder(
+                    borderRadius: BorderRadius.circular(12.0)),
+                child: const Padding(
+                  padding: EdgeInsets.all(16.0),
+                  child: SizedBox.square(
+                      dimension: 40.0, child: LoadingIndictor()),
+                ),
+              ),
+            );
+      },
+    );
     overlayState.insert(overlayEntryOpacity);
     overlayState.insert(overlayEntryLoader);
 
@@ -33,14 +46,13 @@ extension OverlayExtension on DevEssential {
 
     try {
       data = await asyncFunction();
+      overlayEntryLoader.remove();
+      overlayEntryOpacity.remove();
+      return data;
     } on Exception catch (_) {
       overlayEntryLoader.remove();
       overlayEntryOpacity.remove();
       rethrow;
     }
-
-    overlayEntryLoader.remove();
-    overlayEntryOpacity.remove();
-    return data;
   }
 }
