@@ -1,1 +1,9 @@
-export 'splash/splash.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+
+import '../../dev_essentials.dart';
+
+part 'splash/splash_state.dart';
+part 'splash/splash_cubit.dart';
+part 'splash/splash_view.dart';
+part 'root/root_view.dart';
